requirements.txt
flask
anthropic
python-dotenv
PyPDF2
python-docx
jinja2
Werkzeug
